test: split Codex and workflow tests
