+++ test/clawsweeper.test.ts
@@ -1,16 +1,7 @@
 import assert from "node:assert/strict";
 import { execFileSync } from "node:child_process";
-import {
-  chmodSync,
-  existsSync,
-  mkdirSync,
-  mkdtempSync,
-  readFileSync,
-  rmSync,
-  statSync,
-  writeFileSync,
-} from "node:fs";
-import { delimiter, join } from "node:path";
+import { existsSync, mkdirSync, mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import { join } from "node:path";
 import test from "node:test";
 
 import {
@@ -36,7 +27,6 @@ import {
   isProtectedItem,
   itemNumbersArg,
   lockedConversationApplyReason,
-  makeTreeReadOnlyForTest,
   parseGhJson,
   parseGhJsonLines,
   parseDecision,
@@ -51,18 +41,12 @@ import {
   reviewPromptForTest,
   renderReviewCommentFromReport,
   renderReviewContextBudgetForTest,
-  rootCauseClusterFromReportForTest,
-  restoreTreeModesForTest,
   reviewContextLedgerForTest,
   reviewDecisionSchemaText,
   reviewPromptTelemetryForTest,
   reviewPromptTemplate,
-  runCodexForTest,
-  lowerCodexReasoningEffort,
   runtimeBudgetExceeded,
   safeOutputTail,
-  redactInternalCodexModel,
-  codexFailureDecisionForTest,
   shardItemNumbers,
   shouldSyncReviewComment,
   shouldRetryGh,
@@ -4577,2130 +4561,6 @@ test("runtime budget only trips after a positive elapsed limit", () => {
   assert.equal(runtimeBudgetExceeded(1000, 0, 100000), false);
   assert.equal(runtimeBudgetExceeded(1000, 5000, 5999), false);
   assert.equal(runtimeBudgetExceeded(1000, 5000, 6000), true);
-});
-
-test("runCodex accepts valid structured output after non-zero Codex exit", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const outputIndex = process.argv.indexOf("--output-last-message");
-if (outputIndex === -1) process.exit(2);
-fs.writeFileSync(process.argv[outputIndex + 1], process.env.CODEX_DECISION_JSON);
-process.stderr.write("wrote structured output before shutdown failure\\n");
-process.exit(1);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const originalPath = process.env.PATH;
-  const originalDecision = process.env.CODEX_DECISION_JSON;
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_DECISION_JSON = JSON.stringify(
-    closeDecision({
-      decision: "keep_open",
-      closeReason: "none",
-      confidence: "medium",
-      summary: "Keep open for maintainer follow-up.",
-      bestSolution: "Review the routing invariant.",
-      closeComment: "",
-      workReason: "Maintainer review is required.",
-    }),
-  );
-  try {
-    const decision = runCodexForTest({
-      item: item({ number: 83393 }),
-      context: { issue: {}, comments: [], timeline: [] },
-      git: { mainSha: "abc123", latestRelease: null },
-      model: "model-test",
-      openclawDir,
-      reasoningEffort: "high",
-      sandboxMode: "read-only",
-      serviceTier: "",
-      timeoutMs: 10_000,
-      workDir,
-      prompt: "Return a review decision.",
-    });
-
-    assert.equal(decision.decision, "keep_open");
-    assert.equal(decision.summary, "Keep open for maintainer follow-up.");
-  } finally {
-    if (originalPath === undefined) delete process.env.PATH;
-    else process.env.PATH = originalPath;
-    if (originalDecision === undefined) delete process.env.CODEX_DECISION_JSON;
-    else process.env.CODEX_DECISION_JSON = originalDecision;
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex honors env login config unless preserving local Codex auth", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const argsPath = join(root, "codex-args.json");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-fs.writeFileSync(process.env.CODEX_ARGS_PATH, JSON.stringify(process.argv.slice(2)));
-const outputIndex = process.argv.indexOf("--output-last-message");
-if (outputIndex === -1) process.exit(2);
-fs.writeFileSync(process.argv[outputIndex + 1], process.env.CODEX_DECISION_JSON);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ARGS_PATH: process.env.CODEX_ARGS_PATH,
-    CODEX_DECISION_JSON: process.env.CODEX_DECISION_JSON,
-    CLAWSWEEPER_CODEX_LOGIN_METHOD: process.env.CLAWSWEEPER_CODEX_LOGIN_METHOD,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ARGS_PATH = argsPath;
-  process.env.CLAWSWEEPER_CODEX_LOGIN_METHOD = "chatgpt";
-  process.env.CODEX_DECISION_JSON = JSON.stringify(
-    closeDecision({
-      decision: "keep_open",
-      closeReason: "none",
-      confidence: "medium",
-      summary: "Keep open for maintainer follow-up.",
-      bestSolution: "Review the routing invariant.",
-      closeComment: "",
-      workReason: "Maintainer review is required.",
-    }),
-  );
-
-  const runAndReadArgs = (preserveCodexAuth: boolean): string[] => {
-    const decision = runCodexForTest({
-      item: item({ number: 83395 }),
-      context: { issue: {}, comments: [], timeline: [] },
-      git: { mainSha: "abc123", latestRelease: null },
-      model: "model-test",
-      openclawDir,
-      reasoningEffort: "high",
-      sandboxMode: "read-only",
-      serviceTier: "",
-      preserveCodexAuth,
-      timeoutMs: 10_000,
-      workDir,
-      prompt: "Return a review decision.",
-    });
-    assert.equal(decision.decision, "keep_open");
-    return JSON.parse(readFileSync(argsPath, "utf8")) as string[];
-  };
-
-  try {
-    assert.ok(runAndReadArgs(false).includes('forced_login_method="chatgpt"'));
-    assert.equal(runAndReadArgs(true).includes('forced_login_method="chatgpt"'), false);
-    assert.equal(
-      runAndReadArgs(true).some((arg) => arg.startsWith("forced_login_method=")),
-      false,
-    );
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex preserves redacted process output when Codex exits without a decision", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-process.stdout.write("startup banner GH_TOKEN=ghp_abcdefghijklmnopqrstuvwxyz123456 CODEX_ACCESS_TOKEN=codex-access-token-secret\\n");
-process.stderr.write("Rate limit reached for model-test on tokens per min (TPM); OPENAI_API_KEY=sk-proj-abcdefghijklmnopqrstuvwxyz123456 {\\"CODEX_ACCESS_TOKEN\\":\\"codex-json-token-secret\\"}\\n");
-process.exit(1);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const originalPath = process.env.PATH;
-  const originalAttempts = process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS;
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "1";
-  try {
-    assert.throws(
-      () =>
-        runCodexForTest({
-          item: item({ number: 83394 }),
-          context: { issue: {}, comments: [], timeline: [] },
-          git: { mainSha: "abc123", latestRelease: null },
-          model: "model-test",
-          openclawDir,
-          reasoningEffort: "high",
-          sandboxMode: "read-only",
-          serviceTier: "",
-          timeoutMs: 10_000,
-          workDir,
-          prompt: "Return a review decision.",
-        }),
-      (error: unknown) => {
-        const reviewError = error as Error & {
-          status?: number | null;
-          stderr?: string;
-          stdout?: string;
-        };
-        assert.equal(reviewError.status, 1);
-        assert.match(reviewError.stderr ?? "", /Rate limit reached/);
-        assert.match(reviewError.stderr ?? "", /OPENAI_API_KEY=\[REDACTED\]/);
-        assert.match(reviewError.stderr ?? "", /"CODEX_ACCESS_TOKEN":"\[REDACTED\]"/);
-        assert.doesNotMatch(reviewError.stderr ?? "", /sk-proj-/);
-        assert.doesNotMatch(reviewError.stderr ?? "", /codex-json-token-secret/);
-        assert.match(reviewError.stdout ?? "", /startup banner/);
-        assert.match(reviewError.stdout ?? "", /GH_TOKEN=\[REDACTED\]/);
-        assert.match(reviewError.stdout ?? "", /CODEX_ACCESS_TOKEN=\[REDACTED\]/);
-        assert.doesNotMatch(reviewError.stdout ?? "", /ghp_/);
-        assert.doesNotMatch(reviewError.stdout ?? "", /codex-access-token-secret/);
-        return true;
-      },
-    );
-  } finally {
-    if (originalPath === undefined) delete process.env.PATH;
-    else process.env.PATH = originalPath;
-    if (originalAttempts === undefined) delete process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS;
-    else process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = originalAttempts;
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex accepts structured output after more than 128 MiB of process output", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const chunk = Buffer.alloc(1024 * 1024, "x");
-for (let index = 0; index < 129; index += 1) fs.writeSync(1, chunk);
-const outputIndex = process.argv.indexOf("--output-last-message");
-fs.writeFileSync(process.argv[outputIndex + 1], process.env.CODEX_DECISION_JSON);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const originalPath = process.env.PATH;
-  const originalDecision = process.env.CODEX_DECISION_JSON;
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_DECISION_JSON = JSON.stringify(
-    closeDecision({
-      decision: "keep_open",
-      closeReason: "none",
-      confidence: "medium",
-      summary: "Review survived verbose Codex output.",
-      bestSolution: "Keep file-backed process output.",
-      closeComment: "",
-      workReason: "No additional implementation is required.",
-    }),
-  );
-  try {
-    const decision = runCodexForTest({
-      item: item({ number: 83395 }),
-      context: { issue: {}, comments: [], timeline: [] },
-      git: { mainSha: "abc123", latestRelease: null },
-      model: "model-test",
-      openclawDir,
-      reasoningEffort: "high",
-      sandboxMode: "read-only",
-      serviceTier: "",
-      timeoutMs: 20_000,
-      workDir,
-      prompt: "Return a review decision.",
-    });
-
-    assert.equal(decision.summary, "Review survived verbose Codex output.");
-    assert.equal(statSync(join(workDir, "83395.1.codex.stdout.log")).size, 128 * 1024 * 1024);
-  } finally {
-    if (originalPath === undefined) delete process.env.PATH;
-    else process.env.PATH = originalPath;
-    if (originalDecision === undefined) delete process.env.CODEX_DECISION_JSON;
-    else process.env.CODEX_DECISION_JSON = originalDecision;
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("codex failure decisions expose stderr and stdout separately", () => {
-  const errorMessage =
-    "Rate limit reached for model-test on tokens per min (TPM). Please try again in 1ms.";
-  const decision = codexFailureDecisionForTest(
-    1,
-    "Codex review failed for #278 with exit 1.",
-    JSON.stringify({ type: "turn.failed", error: { message: errorMessage } }),
-    "user\nThe reviewed prompt discusses rate limits.",
-  );
-
-  assert.equal(
-    decision.summary,
-    "Codex review failed: retryable codex transport failure (capacity) (exit 1).",
-  );
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "codex stderr")?.detail,
-    "user\nThe reviewed prompt discusses rate limits.",
-  );
-  assert.match(
-    decision.evidence.find((entry) => entry.label === "codex stdout")?.detail ?? "",
-    /"type":"turn.failed"/,
-  );
-});
-
-test("codex failure decisions do not infer buffer overflow from reviewed content", () => {
-  const terminalError =
-    "stream disconnected before completion: The model secret-model-for-test does not exist or you do not have access to it.";
-  const decision = codexFailureDecisionForTest(
-    1,
-    "Codex review failed for #89041 with exit 1.",
-    JSON.stringify({ type: "turn.failed", error: { message: terminalError } }),
-    "user\nThe reviewed PR discusses maxBufferedChunks and maxBuffer behavior.",
-  );
-
-  assert.equal(
-    decision.summary,
-    "Codex review failed: model unavailable or access denied (exit 1).",
-  );
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "codex terminal error")?.detail,
-    terminalError,
-  );
-  assert.equal(decision.codexTerminalFailure, true);
-});
-
-test("codex failure decisions classify structured ENOBUFS as output overflow", () => {
-  const decision = codexFailureDecisionForTest(
-    null,
-    "Codex review failed before producing output.",
-    "",
-    "",
-    { errorCode: "ENOBUFS", signal: "SIGTERM" },
-  );
-
-  assert.equal(decision.summary, "Codex review failed: output buffer overflow.");
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "process error code")?.detail,
-    "ENOBUFS",
-  );
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "process signal")?.detail,
-    "SIGTERM",
-  );
-});
-
-test("codex failure decisions ignore unstructured output and prompt stderr", () => {
-  const decision = codexFailureDecisionForTest(
-    1,
-    "Codex review failed for #92565 with exit 1.",
-    "ERROR: The model quoted-model does not exist or you do not have access to it.",
-    "ERROR: fetch failed",
-  );
-
-  assert.equal(decision.summary, "Codex review failed: codex execution failed (exit 1).");
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "codex terminal error"),
-    undefined,
-  );
-  assert.equal(decision.codexTerminalFailure, false);
-});
-
-test("codex failure decisions trust a final stderr model access denial", () => {
-  const terminalError =
-    "ERROR: stream disconnected before completion: The model secret-model-for-test does not exist or you do not have access to it.";
-  const decision = codexFailureDecisionForTest(
-    1,
-    "Codex review failed for #92565 with exit 1.",
-    "",
-    `reviewed patch text\n${terminalError}`,
-  );
-
-  assert.equal(
-    decision.summary,
-    "Codex review failed: model unavailable or access denied (exit 1).",
-  );
-  assert.equal(
-    decision.evidence.find((entry) => entry.label === "codex terminal error")?.detail,
-    terminalError,
-  );
-  assert.equal(decision.codexTerminalFailure, true);
-});
-
-test("runCodex retries a transient failure in a fresh process", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const codexHome = join(root, "codex-home");
-  const attemptsPath = join(root, "attempts");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  mkdirSync(codexHome, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  writeFileSync(join(codexHome, "config.toml"), 'model = "secret-model-for-test"\n');
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const attemptsPath = process.env.CODEX_ATTEMPTS_PATH;
-const attempt = fs.existsSync(attemptsPath) ? Number(fs.readFileSync(attemptsPath, "utf8")) + 1 : 1;
-fs.writeFileSync(attemptsPath, String(attempt));
-if (attempt === 1) {
-  process.stderr.write("user\\nERROR: The model contributor-quoted-model does not exist or you do not have access to it.\\n");
-  process.stdout.write(JSON.stringify({
-    type: "turn.failed",
-    error: {
-      message: "stream disconnected: Rate limit reached for secret-model-for-test (for limit test) on tokens per min (TPM). Please try again in 1ms."
-    }
-  }) + "\\n");
-  process.exit(1);
-}
-const outputIndex = process.argv.indexOf("--output-last-message");
-fs.writeFileSync(process.argv[outputIndex + 1], process.env.CODEX_DECISION_JSON);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ATTEMPTS_PATH: process.env.CODEX_ATTEMPTS_PATH,
-    CODEX_DECISION_JSON: process.env.CODEX_DECISION_JSON,
-    CODEX_HOME: process.env.CODEX_HOME,
-    CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS: process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS,
-    CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS: process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ATTEMPTS_PATH = attemptsPath;
-  process.env.CODEX_DECISION_JSON = JSON.stringify(
-    closeDecision({
-      decision: "keep_open",
-      closeReason: "none",
-      confidence: "medium",
-      summary: "Review completed after a fresh Codex process.",
-      bestSolution: "Continue the existing review loop.",
-      closeComment: "",
-      workReason: "No additional implementation is required.",
-    }),
-  );
-  process.env.CODEX_HOME = codexHome;
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "2";
-  process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS = "1";
-  try {
-    const decision = runCodexForTest({
-      item: item({ number: 83394 }),
-      context: { issue: {}, comments: [], timeline: [] },
-      git: { mainSha: "abc123", latestRelease: null },
-      model: "internal",
-      openclawDir,
-      reasoningEffort: "high",
-      sandboxMode: "read-only",
-      serviceTier: "",
-      timeoutMs: 10_000,
-      workDir,
-      prompt: "Return a review decision.",
-    });
-
-    assert.equal(readFileSync(attemptsPath, "utf8"), "2");
-    assert.equal(decision.summary, "Review completed after a fresh Codex process.");
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("lowerCodexReasoningEffort steps down one tier and stops at minimal", () => {
-  assert.equal(lowerCodexReasoningEffort("high"), "low");
-  assert.equal(lowerCodexReasoningEffort("HIGH"), "low");
-  assert.equal(lowerCodexReasoningEffort(" medium "), "low");
-  assert.equal(lowerCodexReasoningEffort("low"), "minimal");
-  assert.equal(lowerCodexReasoningEffort("minimal"), null);
-  assert.equal(lowerCodexReasoningEffort("unknown"), null);
-});
-
-test("runCodex completes via a lower-effort fallback after transport exhaustion", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const attemptsPath = join(root, "attempts");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const cfg = process.argv.find((a) => a.startsWith("model_reasoning_effort="));
-const effort = cfg ? cfg.split("=")[1].replace(/"/g, "") : "";
-const attemptsPath = process.env.CODEX_ATTEMPTS_PATH;
-const n = fs.existsSync(attemptsPath) ? Number(fs.readFileSync(attemptsPath, "utf8")) + 1 : 1;
-fs.writeFileSync(attemptsPath, String(n));
-if (effort !== "low") {
-  process.stderr.write("Rate limit reached on tokens per min (TPM). Please try again in 1ms.\\n");
-  process.exit(1);
-}
-const outputIndex = process.argv.indexOf("--output-last-message");
-fs.writeFileSync(process.argv[outputIndex + 1], process.env.CODEX_DECISION_JSON);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ATTEMPTS_PATH: process.env.CODEX_ATTEMPTS_PATH,
-    CODEX_DECISION_JSON: process.env.CODEX_DECISION_JSON,
-    CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS: process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS,
-    CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS: process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS,
-    CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS: process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ATTEMPTS_PATH = attemptsPath;
-  process.env.CODEX_DECISION_JSON = JSON.stringify(
-    closeDecision({
-      decision: "close",
-      closeReason: "duplicate_or_superseded",
-      confidence: "high",
-      summary: "Resolved on main already.",
-      bestSolution: "Close as superseded.",
-      closeComment: "Superseded by main.",
-      workReason: "No additional implementation is required.",
-    }),
-  );
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "2";
-  process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS = "1";
-  process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS = "1";
-  try {
-    const decision = runCodexForTest({
-      item: item({ number: 92181 }),
-      context: { issue: {}, comments: [], timeline: [] },
-      git: { mainSha: "abc123", latestRelease: null },
-      model: "internal",
-      openclawDir,
-      reasoningEffort: "high",
-      sandboxMode: "read-only",
-      serviceTier: "",
-      timeoutMs: 10_000,
-      workDir,
-      prompt: "Return a review decision.",
-    });
-
-    assert.equal(readFileSync(attemptsPath, "utf8"), "3");
-    assert.equal(decision.decision, "close");
-    assert.equal(decision.confidence, "medium");
-    assert.match(decision.summary, /^Degraded review:/);
-    assert.match(decision.summary, /lower-effort \(low\) fallback pass/);
-    assert.match(decision.summary, /Resolved on main already\./);
-    assert.equal(decision.evidence[0]?.label, "degraded review mode");
-    assert.match(decision.evidence[0]?.detail ?? "", /high → low reasoning effort fallback/);
-    assert.equal(decision.evidence[1]?.label, "original codex transport failure");
-    assert.match(decision.evidence[1]?.detail ?? "", /Rate limit reached|tokens per min/i);
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex keeps the transport classification when the fallback also fails", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const attemptsPath = join(root, "attempts");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const attemptsPath = process.env.CODEX_ATTEMPTS_PATH;
-const n = fs.existsSync(attemptsPath) ? Number(fs.readFileSync(attemptsPath, "utf8")) + 1 : 1;
-fs.writeFileSync(attemptsPath, String(n));
-process.stderr.write("Rate limit reached on tokens per min (TPM). Please try again in 1ms.\\n");
-process.exit(1);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ATTEMPTS_PATH: process.env.CODEX_ATTEMPTS_PATH,
-    CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS: process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS,
-    CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS: process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS,
-    CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS: process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ATTEMPTS_PATH = attemptsPath;
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "2";
-  process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS = "1";
-  process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS = "1";
-  try {
-    assert.throws(
-      () =>
-        runCodexForTest({
-          item: item({ number: 92181 }),
-          context: { issue: {}, comments: [], timeline: [] },
-          git: { mainSha: "abc123", latestRelease: null },
-          model: "internal",
-          openclawDir,
-          reasoningEffort: "high",
-          sandboxMode: "read-only",
-          serviceTier: "",
-          timeoutMs: 10_000,
-          workDir,
-          prompt: "Return a review decision.",
-        }),
-      (error: unknown) => {
-        const reviewError = error as Error;
-        assert.equal(readFileSync(attemptsPath, "utf8"), "3");
-        assert.match(reviewError.message, /Lower-effort \(low\) fallback also failed/);
-        const failure = codexFailureDecisionForTest(
-          1,
-          reviewError.message,
-          (reviewError as { stdout?: string }).stdout ?? "",
-          (reviewError as { stderr?: string }).stderr ?? "",
-        );
-        assert.match(failure.summary, /retryable codex transport failure \(capacity\)/);
-        return true;
-      },
-    );
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex skips the lower-effort fallback when the time budget is too small", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const attemptsPath = join(root, "attempts");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const attemptsPath = process.env.CODEX_ATTEMPTS_PATH;
-const n = fs.existsSync(attemptsPath) ? Number(fs.readFileSync(attemptsPath, "utf8")) + 1 : 1;
-fs.writeFileSync(attemptsPath, String(n));
-process.stderr.write("Rate limit reached on tokens per min (TPM). Please try again in 1ms.\\n");
-process.exit(1);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ATTEMPTS_PATH: process.env.CODEX_ATTEMPTS_PATH,
-    CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS: process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS,
-    CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS: process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS,
-    CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS: process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ATTEMPTS_PATH = attemptsPath;
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "2";
-  process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS = "1";
-  process.env.CLAWSWEEPER_CODEX_FALLBACK_MIN_BUDGET_MS = "10000000";
-  try {
-    assert.throws(() =>
-      runCodexForTest({
-        item: item({ number: 92181 }),
-        context: { issue: {}, comments: [], timeline: [] },
-        git: { mainSha: "abc123", latestRelease: null },
-        model: "internal",
-        openclawDir,
-        reasoningEffort: "high",
-        sandboxMode: "read-only",
-        serviceTier: "",
-        timeoutMs: 10_000,
-        workDir,
-        prompt: "Return a review decision.",
-      }),
-    );
-    assert.equal(readFileSync(attemptsPath, "utf8"), "2");
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("runCodex does not retry terminal model access failures", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const openclawDir = join(root, "openclaw");
-  const workDir = join(root, "codex-work");
-  const binDir = join(root, "bin");
-  const attemptsPath = join(root, "attempts");
-  mkdirSync(openclawDir, { recursive: true });
-  mkdirSync(binDir, { recursive: true });
-  execFileSync("git", ["init"], { cwd: openclawDir, stdio: "ignore" });
-  const codexPath = join(binDir, "codex");
-  writeFileSync(
-    codexPath,
-    `#!/usr/bin/env node
-const fs = require("node:fs");
-const attemptsPath = process.env.CODEX_ATTEMPTS_PATH;
-const attempt = fs.existsSync(attemptsPath) ? Number(fs.readFileSync(attemptsPath, "utf8")) + 1 : 1;
-fs.writeFileSync(attemptsPath, String(attempt));
-process.stderr.write("reviewed patch text\\n");
-process.stderr.write("stream disconnected before completion: The model secret-model-for-test does not exist or you do not have access to it.\\n");
-process.exit(1);
-`,
-  );
-  chmodSync(codexPath, 0o755);
-  const previous = {
-    PATH: process.env.PATH,
-    CODEX_ATTEMPTS_PATH: process.env.CODEX_ATTEMPTS_PATH,
-    CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS: process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS,
-    CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS: process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS,
-  };
-  process.env.PATH = `${binDir}${delimiter}${process.env.PATH ?? ""}`;
-  process.env.CODEX_ATTEMPTS_PATH = attemptsPath;
-  process.env.CLAWSWEEPER_CODEX_REVIEW_ATTEMPTS = "3";
-  process.env.CLAWSWEEPER_CODEX_REVIEW_RETRY_DELAY_MS = "1";
-  try {
-    assert.throws(
-      () =>
-        runCodexForTest({
-          item: item({ number: 89041 }),
-          context: { issue: {}, comments: [], timeline: [] },
-          git: { mainSha: "abc123", latestRelease: null },
-          model: "internal",
-          openclawDir,
-          reasoningEffort: "high",
-          sandboxMode: "read-only",
-          serviceTier: "",
-          timeoutMs: 10_000,
-          workDir,
-          prompt: "Return a review decision.",
-        }),
-      (error: unknown) => {
-        const reviewError = error as Error & { stderr?: string };
-        assert.match(reviewError.stderr ?? "", /does not exist or you do not have access/);
-        return true;
-      },
-    );
-    assert.equal(readFileSync(attemptsPath, "utf8"), "1");
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("Codex failure redaction hides the configured internal model", () => {
-  const root = mkdtempSync(tmpPrefix);
-  writeFileSync(join(root, "config.toml"), 'model = "secret-model-for-test"\n');
-  try {
-    const redacted = redactInternalCodexModel(
-      "selected secret-model-for-test; Rate limit reached for unknown-model (for limit test)",
-      root,
-    );
-    assert.doesNotMatch(redacted, /secret-model-for-test|unknown-model/);
-    assert.equal(redacted.match(/\[REDACTED_INTERNAL_MODEL\]/g)?.length, 2);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("Codex failure redaction reads the default home configuration", () => {
-  const root = mkdtempSync(tmpPrefix);
-  const codexHome = join(root, ".codex");
-  mkdirSync(codexHome, { recursive: true });
-  writeFileSync(join(codexHome, "config.toml"), 'model = "default-secret-model"\n');
-  const previous = {
-    HOME: process.env.HOME,
-    CODEX_HOME: process.env.CODEX_HOME,
-    CLAWSWEEPER_INTERNAL_MODEL: process.env.CLAWSWEEPER_INTERNAL_MODEL,
-  };
-  try {
-    process.env.HOME = root;
-    delete process.env.CODEX_HOME;
-    delete process.env.CLAWSWEEPER_INTERNAL_MODEL;
-    assert.equal(
-      redactInternalCodexModel("selected default-secret-model"),
-      "selected [REDACTED_INTERNAL_MODEL]",
-    );
-  } finally {
-    for (const [key, value] of Object.entries(previous)) {
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("decision parser enforces required schema-shaped evidence", () => {
-  assert.equal(parseDecision(closeDecision()).decision, "close");
-  assert.equal(parseDecision(closeDecision({ itemCategory: "skill" })).itemCategory, "skill");
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        evidence: [{ label: "partial", detail: "missing nullable fields" }],
-      }),
-    /decision\.evidence\[0\]\.file/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        likelyOwners: [],
-      }),
-    /decision\.likelyOwners must not be empty/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        likelyOwners: [{ person: "@alice", reason: "missing fields" }],
-      }),
-    /decision\.likelyOwners\[0\]\.role/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        workCandidate: "auto_everything",
-      }),
-    /decision\.workCandidate/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        itemCategory: "mixed_mode",
-      }),
-    /decision\.itemCategory/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        triagePriority: "urgent",
-      }),
-    /decision\.triagePriority/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        impactLabels: ["impact:unknown"],
-      }),
-    /decision\.impactLabels\[0\]/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        impactLabels: [
-          "impact:data-loss",
-          "impact:security",
-          "impact:crash-loop",
-          "impact:message-loss",
-        ],
-      }),
-    /decision\.impactLabels must contain at most 3 labels/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        impactLabels: ["impact:data-loss", "impact:data-loss"],
-      }),
-    /decision\.impactLabels must not contain duplicates/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk:unknown"],
-      }),
-    /decision\.mergeRiskLabels\[0\]/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: [
-          "merge-risk: 🚨 compatibility",
-          "merge-risk: 🚨 message-delivery",
-          "merge-risk: 🚨 session-state",
-          "merge-risk: 🚨 auth-provider",
-        ],
-      }),
-    /decision\.mergeRiskLabels must contain at most 3 labels/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk: 🚨 compatibility", "merge-risk: 🚨 compatibility"],
-      }),
-    /decision\.mergeRiskLabels must not contain duplicates/,
-  );
-  assert.equal(
-    parseDecision({
-      ...closeDecision(),
-      mergeRiskOptions: undefined,
-    }).mergeRiskOptions.length,
-    0,
-  );
-  assert.deepEqual(
-    parseDecision({
-      ...closeDecision(),
-      reviewMetrics: [
-        {
-          label: "Files affected",
-          value: "3 files affected",
-          reason: "The PR touches enough files that maintainers should scan the changed surface.",
-        },
-      ],
-    }).reviewMetrics,
-    [
-      {
-        label: "Files affected",
-        value: "3 files affected",
-        reason: "The PR touches enough files that maintainers should scan the changed surface.",
-      },
-    ],
-  );
-  assert.throws(() => {
-    const decision = closeDecision();
-    delete decision.reviewMetrics;
-    return parseDecision(decision);
-  }, /decision\.reviewMetrics must be an array/);
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        reviewMetrics: [{ label: "Files affected", value: "3 files affected" }],
-      }),
-    /decision\.reviewMetrics\[0\]\.reason/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskOptions: [
-          {
-            title: "Accept the risk",
-            body: "Merge only if maintainers accept this risk.",
-            category: "accept_risk",
-            recommended: false,
-            automergeInstruction: "",
-          },
-        ],
-      }),
-    /decision\.mergeRiskOptions must be empty when mergeRiskLabels is empty/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk: 🚨 compatibility"],
-      }),
-    /decision\.mergeRiskOptions must include 1-3 options when mergeRiskLabels is not empty/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk: 🚨 compatibility"],
-        mergeRiskOptions: [
-          {
-            title: "Preserve behavior",
-            body: "Keep the existing default behavior before merge.",
-            category: "fix_before_merge",
-            recommended: true,
-            automergeInstruction: "Keep the existing default behavior before merge.",
-          },
-          {
-            title: "Accept risk",
-            body: "Merge only if maintainers accept the compatibility break.",
-            category: "accept_risk",
-            recommended: true,
-            automergeInstruction: "",
-          },
-        ],
-      }),
-    /decision\.mergeRiskOptions must not contain more than one recommended option/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk: 🚨 security-boundary"],
-        mergeRiskOptions: [
-          {
-            title: "Accept risk",
-            body: "Merge only if maintainers accept the hardening tradeoff.",
-            category: "accept_risk",
-            recommended: true,
-            automergeInstruction: "Merge the intentional hardening change.",
-          },
-        ],
-      }),
-    /decision\.mergeRiskOptions\[0\]\.automergeInstruction requires fix_before_merge category/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        mergeRiskLabels: ["merge-risk: 🚨 message-delivery"],
-        mergeRiskOptions: [
-          {
-            title: "Guard delivery",
-            body: "Add delivery-state tests before merge.",
-            category: "fix_before_merge",
-            recommended: false,
-            automergeInstruction: "Add delivery-state tests before merge.",
-          },
-        ],
-      }),
-    /decision\.mergeRiskOptions\[0\]\.automergeInstruction requires a recommended option/,
-  );
-  assert.deepEqual(
-    parseDecision(
-      closeDecision({
-        impactLabels: ["impact:other"],
-        labelJustifications: [
-          {
-            label: "P2",
-            reason: "Normal priority applies to this limited-scope implemented behavior check.",
-          },
-          {
-            label: "impact:other",
-            reason: "The issue has maintainer-visible impact outside the specific taxonomy.",
-          },
-        ],
-      }),
-    ).impactLabels,
-    ["impact:other"],
-  );
-  assert.deepEqual(
-    parseDecision(
-      closeDecision({
-        mergeRiskLabels: ["merge-risk: 🚨 other"],
-        mergeRiskOptions: [
-          {
-            title: "Validate the uncategorized risk",
-            body: "Run targeted validation for the maintainer-visible risk before merge.",
-            category: "fix_before_merge",
-            recommended: true,
-            automergeInstruction:
-              "Run targeted validation for the maintainer-visible risk before merge.",
-          },
-        ],
-        labelJustifications: [
-          {
-            label: "P2",
-            reason: "Normal priority applies to this limited-scope implemented behavior check.",
-          },
-          {
-            label: "merge-risk: 🚨 other",
-            reason: "The PR has a maintainer-visible merge risk outside the specific taxonomy.",
-          },
-        ],
-      }),
-    ).mergeRiskLabels,
-    ["merge-risk: 🚨 other"],
-  );
-  assert.throws(() => {
-    const decision = closeDecision();
-    delete decision.labelJustifications;
-    return parseDecision(decision);
-  }, /decision\.labelJustifications must be an array/);
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision({
-          impactLabels: ["impact:message-loss"],
-          labelJustifications: [
-            {
-              label: "P2",
-              reason: "Normal priority applies to this limited-scope implemented behavior check.",
-            },
-          ],
-        }),
-      }),
-    /decision\.labelJustifications missing selected labels: impact:message-loss/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision({
-          labelJustifications: [
-            {
-              label: "P2",
-              reason: "Normal priority applies to this limited-scope implemented behavior check.",
-            },
-            {
-              label: "impact:data-loss",
-              reason: "The selected labels did not include this impact area.",
-            },
-          ],
-        }),
-      }),
-    /decision\.labelJustifications contains unselected labels: impact:data-loss/,
-  );
-  assert.throws(
-    () =>
-      parseDecision({
-        ...closeDecision(),
-        requiresNewConfigOption: "false",
-      }),
-    /decision\.requiresNewConfigOption/,
-  );
-  assert.throws(() => {
-    const decision = closeDecision();
-    delete decision.securityReview;
-    return parseDecision(decision);
-  }, /decision\.securityReview/);
-  assert.throws(() => {
-    const decision = closeDecision();
-    delete decision.realBehaviorProof;
-    return parseDecision(decision);
-  }, /decision\.realBehaviorProof/);
-  const workCandidate = parseDecision(
-    closeDecision({
-      decision: "keep_open",
-      closeReason: "none",
-      confidence: "medium",
-      workCandidate: "queue_fix_pr",
-      workConfidence: "high",
-      workPriority: "medium",
-      workReason: "The bug is narrow and reproducible.",
-      workPrompt: "Fix the narrow bug and add a regression test.",
-      workClusterRefs: ["#123", "#456"],
-      workValidation: ["pnpm test:unit"],
-      workLikelyFiles: ["src/example.ts", "test/example.test.ts"],
-    }),
-  );
-  assert.equal(workCandidate.workCandidate, "queue_fix_pr");
-  assert.equal(workCandidate.triagePriority, "P2");
-  assert.equal(workCandidate.itemCategory, "bug");
-  assert.equal(workCandidate.reproductionStatus, "reproduced");
-  assert.equal(workCandidate.realBehaviorProof.status, "not_applicable");
-  assert.deepEqual(workCandidate.workClusterRefs, ["#123", "#456"]);
-});
-
-test("decision parser validates typed root-cause clusters", () => {
-  const canonicalRef = "https://github.com/openclaw/openclaw/pull/456";
-  const canonicalIssueRef = "https://github.com/openclaw/openclaw/issues/456";
-  const candidatePullRef = "https://github.com/openclaw/openclaw/pull/789";
-  const independentRootCauseCluster = {
-    confidence: "low",
-    canonicalRef: null,
-    currentItemRelationship: "independent",
-    summary: "No evidence-backed root-cause cluster was established.",
-    members: [],
-  };
-  const rootCauseCluster = {
-    confidence: "high",
-    canonicalRef,
-    currentItemRelationship: "fixed_by_candidate",
-    summary: "The candidate PR fixes the reproduced issue.",
-    members: [
-      {
-        ref: canonicalRef,
-        relationship: "canonical",
-        reason: "The PR contains the focused fix and regression test.",
-      },
-    ],
-  };
-  const parsed = parseDecision(
-    closeDecision({ rootCauseCluster }),
-    item({ repo: "openclaw/openclaw", number: 123, kind: "issue" }),
-  );
-  assert.deepEqual(parsed.rootCauseCluster, rootCauseCluster);
-
-  const prCandidateForCanonicalIssue = {
-    confidence: "high",
-    canonicalRef: canonicalIssueRef,
-    currentItemRelationship: "fixed_by_candidate",
-    summary: "This PR is the candidate fix for the canonical issue.",
-    members: [
-      {
-        ref: canonicalIssueRef,
-        relationship: "canonical",
-        reason: "The issue tracks the underlying user-visible bug.",
-      },
-    ],
-  };
-  assert.deepEqual(
-    parseDecision(
-      closeDecision({ rootCauseCluster: prCandidateForCanonicalIssue }),
-      item({ kind: "pull_request" }),
-    ).rootCauseCluster,
-    prCandidateForCanonicalIssue,
-  );
-
-  const canonicalIssueWithCandidateMember = {
-    confidence: "high",
-    canonicalRef: "https://github.com/openclaw/openclaw/issues/123",
-    currentItemRelationship: "canonical",
-    summary: "The issue is canonical and has an open candidate fix PR.",
-    members: [
-      {
-        ref: candidatePullRef,
-        relationship: "fixed_by_candidate",
-        reason: "The PR carries the candidate fix for this canonical issue.",
-      },
-    ],
-  };
-  assert.deepEqual(
-    parseDecision(closeDecision({ rootCauseCluster: canonicalIssueWithCandidateMember }), item())
-      .rootCauseCluster,
-    canonicalIssueWithCandidateMember,
-  );
-
-  const invalidRootCauseClusters = [
-    {
-      ...rootCauseCluster,
-      members: [...rootCauseCluster.members, ...rootCauseCluster.members],
-    },
-    {
-      ...rootCauseCluster,
-      canonicalRef: "https://github.com/other/repo/pull/456",
-      members: [
-        {
-          ...rootCauseCluster.members[0],
-          ref: "https://github.com/other/repo/pull/456",
-        },
-      ],
-    },
-    {
-      ...rootCauseCluster,
-      members: [
-        ...rootCauseCluster.members,
-        {
-          ref: "https://github.com/openclaw/openclaw/issues/789",
-          relationship: "canonical",
-          reason: "A conflicting second canonical item.",
-        },
-      ],
-    },
-    {
-      ...rootCauseCluster,
-      members: [
-        {
-          ...rootCauseCluster.members[0],
-          ref: "https://github.com/openclaw/openclaw/pull/789",
-        },
-      ],
-    },
-    {
-      ...rootCauseCluster,
-      canonicalRef: canonicalIssueRef,
-      members: [
-        {
-          ...rootCauseCluster.members[0],
-          ref: canonicalIssueRef,
-        },
-      ],
-    },
-    {
-      ...canonicalIssueWithCandidateMember,
-      members: [
-        {
-          ref: "https://github.com/openclaw/openclaw/issues/789",
-          relationship: "fixed_by_candidate",
-          reason: "Issue-to-issue candidate-fix labels are not meaningful.",
-        },
-      ],
-    },
-    {
-      ...rootCauseCluster,
-      members: [
-        {
-          ref: "https://github.com/openclaw/openclaw/issues/123",
-          relationship: "canonical",
-          reason: "Incorrectly repeats the current item.",
-        },
-      ],
-      canonicalRef: "https://github.com/openclaw/openclaw/issues/123",
-      currentItemRelationship: "duplicate",
-    },
-    {
-      ...rootCauseCluster,
-      members: [
-        {
-          ref: "https://github.com/OpenClaw/OpenClaw/issues/123",
-          relationship: "canonical",
-          reason: "Incorrectly repeats the current item with different casing.",
-        },
-      ],
-      canonicalRef: "https://github.com/OpenClaw/OpenClaw/issues/123",
-      currentItemRelationship: "duplicate",
-    },
-    {
-      ...rootCauseCluster,
-      members: [
-        rootCauseCluster.members[0],
-        {
-          ...rootCauseCluster.members[0],
-          ref: "https://github.com/OpenClaw/OpenClaw/pull/456",
-        },
-      ],
-    },
-  ];
-
-  for (const invalidRootCauseCluster of invalidRootCauseClusters) {
-    assert.deepEqual(
-      parseDecision(
-        closeDecision({
-          rootCauseCluster: invalidRootCauseCluster,
-        }),
-        item(),
-      ).rootCauseCluster,
-      independentRootCauseCluster,
-    );
-  }
-});
-
-test("root-cause report parsing defaults legacy and malformed reports safely", () => {
-  assert.deepEqual(rootCauseClusterFromReportForTest(reportFrontMatter({ number: "123" })), {
-    confidence: "low",
-    canonicalRef: null,
-    currentItemRelationship: "independent",
-    summary: "No evidence-backed root-cause cluster was established.",
-    members: [],
-  });
-  assert.deepEqual(
-    rootCauseClusterFromReportForTest(
-      reportFrontMatter({
-        number: "123",
-        root_cause_cluster: "{not-json",
-      }),
-    ),
-    {
-      confidence: "low",
-      canonicalRef: null,
-      currentItemRelationship: "independent",
-      summary: "No evidence-backed root-cause cluster was established.",
-      members: [],
-    },
-  );
-  const valid = {
-    confidence: "high",
-    canonicalRef: "https://github.com/openclaw/openclaw/issues/456",
-    currentItemRelationship: "duplicate",
-    summary: "The other issue is the canonical report.",
-    members: [
-      {
-        ref: "https://github.com/openclaw/openclaw/issues/456",
-        relationship: "canonical",
-        reason: "It has the complete reproduction and accepted scope.",
-      },
-    ],
-  };
-  assert.deepEqual(
-    rootCauseClusterFromReportForTest(
-      reportFrontMatter({
-        number: "123",
-        root_cause_cluster: JSON.stringify(valid),
-      }),
-    ),
-    valid,
-  );
-});
-
-test("review workflow gives Codex a read-only inspection token", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const eventReviewJobStart = workflow.indexOf("\n  event-review-apply:");
-  const planJobStart = workflow.indexOf("\n  plan:", eventReviewJobStart);
-  const eventReviewJob = workflow.slice(eventReviewJobStart, planJobStart);
-  const reviewJobStart = workflow.indexOf("\n  review:");
-  const publishJobStart = workflow.indexOf("\n  publish:", reviewJobStart);
-  const reviewJob = workflow.slice(reviewJobStart, publishJobStart);
-  const exactReviewStart = eventReviewJob.indexOf("- name: Review exact event item");
-  const stateTokenStart = eventReviewJob.indexOf("- name: Create state token", exactReviewStart);
-  const exactReviewStep = eventReviewJob.slice(exactReviewStart, stateTokenStart);
-
-  assert.match(workflow, /id: codex-inspection-token/);
-  assert.match(workflow, /permission-issues: read/);
-  assert.match(workflow, /CLAWSWEEPER_PROOF_INSPECTION_TOKEN/);
-  assert.match(
-    exactReviewStep,
-    /CLAWSWEEPER_PROOF_INSPECTION_TOKEN: \$\{\{ steps\.target-read-token\.outputs\.token \|\| github\.token \}\}/,
-  );
-  assert.match(reviewJob, /uses: \.\/clawsweeper\/\.github\/actions\/setup-codex/);
-  assert.doesNotMatch(reviewJob, /uses: \.\/\.github\/actions\/setup-codex/);
-});
-
-test("dashboard syncs Worker secrets with durable lifecycle storage", () => {
-  const workflow = readFileSync(".github/workflows/dashboard.yml", "utf8");
-  const config = readFileSync("dashboard/wrangler.toml", "utf8");
-
-  assert.doesNotMatch(workflow, /storage\/kv\/namespaces/);
-  assert.match(config, /\[\[durable_objects\.bindings\]\]/);
-  assert.match(config, /name = "STATUS_STORE"/);
-  assert.match(config, /class_name = "StatusStore"/);
-  assert.match(config, /new_sqlite_classes = \["StatusStore"\]/);
-  assert.match(workflow, /workers\/scripts\/\$CLOUDFLARE_WORKER_NAME\/secrets-bulk/);
-  assert.match(workflow, /Content-Type: application\/merge-patch\+json/);
-  assert.match(workflow, /jq -e '\.success == true'/);
-  assert.doesNotMatch(workflow, /wrangler@4\.90\.0 secret bulk/);
-});
-
-test("publish workflow installs Codex from the root checkout path", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const publishJobStart = workflow.indexOf("\n  publish:");
-  const recoverJobStart = workflow.indexOf("\n  recover-review-failures:", publishJobStart);
-  const publishJob = workflow.slice(publishJobStart, recoverJobStart);
-
-  assert.match(publishJob, /uses: \.\/\.github\/actions\/setup-codex/);
-  assert.doesNotMatch(publishJob, /uses: \.\/clawsweeper\/\.github\/actions\/setup-codex/);
-  const setupCodexStart = publishJob.indexOf("- uses: ./.github/actions/setup-codex");
-  const syncCommentsStart = publishJob.indexOf("- name: Sync selected review comments");
-  const applySelectedStart = publishJob.indexOf("- name: Apply selected safe close proposals");
-  assert.ok(setupCodexStart > syncCommentsStart);
-  assert.ok(applySelectedStart > setupCodexStart);
-  assert.match(
-    publishJob.slice(setupCodexStart, applySelectedStart),
-    /if: \$\{\{ success\(\) && steps\.target-write-token\.outputs\.token != '' && github\.event\.inputs\.apply_after_review == 'true' \}\}/,
-  );
-});
-
-test("apply workflow installs Codex only when proof-eligible apply work can run", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8").replace(/\r\n/g, "\n");
-  const applyJobStart = workflow.indexOf("\n  apply-existing:");
-  assert.notEqual(applyJobStart, -1);
-  const applyJob = workflow.slice(applyJobStart);
-  const reconcileStart = applyJob.indexOf("- name: Reconcile before apply preselect");
-  const preselectStart = applyJob.indexOf("- name: Preselect apply work that can need Codex");
-  const setupCodexStart = applyJob.indexOf("- uses: ./.github/actions/setup-codex", preselectStart);
-  const applyStart = applyJob.indexOf(
-    "- name: Apply unchanged proposed decisions with checkpoints",
-  );
-
-  assert.ok(reconcileStart !== -1);
-  assert.ok(preselectStart !== -1);
-  assert.ok(preselectStart > reconcileStart);
-  assert.ok(setupCodexStart > preselectStart);
-  assert.ok(applyStart > setupCodexStart);
-  const reconcileBlock = applyJob.slice(reconcileStart, preselectStart);
-  assert.match(reconcileBlock, /GH_TOKEN: \$\{\{ steps\.target-write-token\.outputs\.token \}\}/);
-  assert.match(reconcileBlock, /pnpm run reconcile -- "\$\{reconcile_args\[@\]\}"/);
-  assert.match(
-    applyJob.slice(setupCodexStart, applyStart),
-    /if: \$\{\{ steps\.apply-preselect\.outputs\.needs_codex == 'true' \}\}/,
-  );
-  const preselectBlock = applyJob.slice(preselectStart, setupCodexStart);
-  assert.match(preselectBlock, /\[ "\$sync_comments_only" = "true" \]/);
-  assert.match(preselectBlock, /comment-sync-batch/);
-  assert.match(preselectBlock, /batch_count="\$\(awk -F=/);
-  const syncOnlyStart = preselectBlock.indexOf('if [ "$sync_comments_only" = "true" ]; then');
-  assert.ok(syncOnlyStart !== -1);
-  const nonSyncMatch = /\n\s+else\n\s+proof_args=\(/.exec(preselectBlock.slice(syncOnlyStart));
-  assert.ok(nonSyncMatch);
-  const nonSyncStart = syncOnlyStart + nonSyncMatch.index;
-  assert.ok(nonSyncStart > syncOnlyStart);
-  assert.doesNotMatch(preselectBlock.slice(syncOnlyStart, nonSyncStart), /needs_codex=true/);
-  assert.match(preselectBlock, /\[ -n "\$item_numbers" \]/);
-  assert.match(preselectBlock, /proposed-pr-close-coverage-item-numbers/);
-  assert.match(preselectBlock, /proof_args\+=\(--item-numbers "\$item_numbers"\)/);
-  assert.match(preselectBlock, /if \[ -n "\$selected" \]; then\s+needs_codex=true/);
-  assert.doesNotMatch(preselectBlock, /if \[ -n "\$item_numbers" \]; then\s+needs_codex=true/);
-  assert.doesNotMatch(preselectBlock, /normalized_apply_close_reasons=/);
-});
-
-test("apply workflow bounds checkpoints and requeues with a fresh token", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8").replace(/\r\n/g, "\n");
-  const inputBlock = workflow.slice(
-    workflow.indexOf("  workflow_dispatch:\n    inputs:"),
-    workflow.indexOf("\n  schedule:"),
-  );
-  const applyJob = workflow.slice(workflow.indexOf("\n  apply-existing:"));
-  const applyStep = applyJob.slice(
-    applyJob.indexOf("- name: Apply unchanged proposed decisions with checkpoints"),
-    applyJob.indexOf("- name: Commit apply results"),
-  );
-  const continueStep = applyJob.slice(
-    applyJob.indexOf("- name: Continue apply sweep"),
-    applyJob.indexOf("- name: Queue review backstops"),
-  );
-
-  assert.match(inputBlock, /apply_limit:[\s\S]*default: "5"/);
-  assert.match(inputBlock, /apply_checkpoint_size:[\s\S]*default: "5"/);
-  assert.match(applyStep, /Capping apply checkpoint size at 5/);
-  assert.match(applyStep, /close_processed_limit=300/);
-  assert.match(applyStep, /processed-limit "\$close_processed_limit"/);
-  assert.match(applyStep, /comment_sync_processed_limit=1000/);
-  assert.match(applyStep, /--processed-limit "\$comment_sync_processed_limit"/);
-  assert.match(applyStep, /reached its \$close_processed_limit-record budget/);
-  assert.match(applyStep, /apply_close_reasons="\$\(printf '%s\\n' "\$apply_close_reasons"/);
-  assert.match(applyStep, /No enabled close reasons remain after policy filtering/);
-  assert.match(applyStep, /true\|1\|yes\|on\) product_direction_enabled=true/);
-  assert.match(
-    applyStep,
-    /if \[ "\$result_count" -ge "\$close_processed_limit" \] && \[ "\$closed_in_chunk" -gt 0 \]/,
-  );
-  assert.match(applyStep, /sync_comments_only" != "true" .*apply_close_reasons/);
-  assert.match(applyStep, /continue_apply=true/);
-  assert.match(applyStep, /break\n\s+done/);
-  assert.match(applyStep, /echo "APPLY_CONTINUE=\$continue_apply"/);
-  assert.match(continueStep, /APPLY_CONTINUE:-false/);
-  assert.doesNotMatch(continueStep, /APPLY_CLOSED_TOTAL:-0.*APPLY_LIMIT:-0/);
-});
-
-test("apply workflow syncs source checkout before state hydration", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8").replace(/\r\n/g, "\n");
-  const applyJobStart = workflow.indexOf("\n  apply-existing:");
-  assert.notEqual(applyJobStart, -1);
-  const applyJob = workflow.slice(applyJobStart);
-  const resolveTargetStart = applyJob.indexOf("- name: Resolve target repository");
-  const syncStart = applyJob.indexOf("- name: Sync source checkout before state hydration");
-  const setupStateStart = applyJob.indexOf("- uses: ./.github/actions/setup-state");
-  const reconcileStart = applyJob.indexOf("- name: Reconcile before apply preselect");
-
-  assert.ok(resolveTargetStart !== -1);
-  assert.ok(syncStart > resolveTargetStart);
-  assert.ok(setupStateStart > syncStart);
-  assert.ok(reconcileStart > setupStateStart);
-  assert.equal(applyJob.indexOf("- name: Sync before applying decisions"), -1);
-  assert.match(applyJob.slice(syncStart, setupStateStart), /run: git pull --rebase/);
-  assert.doesNotMatch(applyJob.slice(setupStateStart, reconcileStart), /git pull --rebase/);
-});
-
-test("sweep target tokens fall back when an org app installation is missing", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const stepBlocks = (name: string) =>
-    workflow
-      .split(`- name: ${name}`)
-      .slice(1)
-      .map((block) => block.split("\n      - ")[0]);
-
-  assert.match(
-    workflow,
-    /CLAWSWEEPER_INVENTORY_TOKEN_STEIPETE: \$\{\{ steps\.steipete-token\.outputs\.token \|\| '__public__' \}\}/,
-  );
-  const openclawInventoryBlocks = stepBlocks("Create OpenClaw inventory token");
-  assert.equal(openclawInventoryBlocks.length, 1);
-  assert.doesNotMatch(openclawInventoryBlocks[0] ?? "", /continue-on-error: true/);
-  for (const name of [
-    "Create target read token",
-    "Create target write token",
-    "Create target review token",
-    "Create target Codex inspection token",
-  ]) {
-    const blocks = stepBlocks(name);
-    assert.ok(blocks.length > 0, `missing workflow step: ${name}`);
-    for (const block of blocks) {
-      assert.match(block, /continue-on-error: true/);
-    }
-  }
-  assert.match(
-    workflow,
-    /GH_TOKEN: \$\{\{ steps\.target-read-token\.outputs\.token \|\| github\.token \}\}/,
-  );
-  assert.match(
-    workflow,
-    /CLAWSWEEPER_PROOF_INSPECTION_TOKEN: \$\{\{ steps\.codex-inspection-token\.outputs\.token \|\| github\.token \}\}/,
-  );
-  assert.ok(
-    workflow.includes(
-      "if: ${{ success() && steps.target-write-token.outputs.token != '' && needs.plan.outputs.hot_intake != 'true'",
-    ),
-  );
-  assert.ok(
-    workflow.includes(
-      "if: ${{ success() && steps.target-write-token.outputs.token != '' && ((github.event_name == 'repository_dispatch'",
-    ),
-  );
-  assert.ok(
-    workflow.includes(
-      "if: ${{ success() && steps.target-write-token.outputs.token != '' && github.event.inputs.apply_after_review == 'true' }}",
-    ),
-  );
-  assert.doesNotMatch(workflow, new RegExp("OPENCLAW_" + "GH_TOKEN"));
-});
-
-test("proof nudge workflow is manual-first and scheduled behind repo vars", () => {
-  const sweepWorkflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const workflow = readFileSync(".github/workflows/proof-nudges.yml", "utf8");
-  const job = workflow.slice(workflow.indexOf("  proof-nudges:"), workflow.length);
-  const concurrency = workflow.slice(workflow.indexOf("concurrency:"), workflow.indexOf("\njobs:"));
-
-  assert.doesNotMatch(sweepWorkflow, /proof_nudges/);
-  assert.match(workflow, /execute:[\s\S]*?default: "false"/);
-  assert.match(workflow, /cron: "0 10 \* \* \*"/);
-  assert.doesNotMatch(workflow, /cron: "0 11 \* \* \*"/);
-  assert.match(concurrency, /clawsweeper-proof-nudges/);
-  assert.doesNotMatch(job, /Check scheduled Central time/);
-  assert.doesNotMatch(job, /PROOF_NUDGES_SCHEDULE_TZ/);
-  assert.doesNotMatch(job, /PROOF_NUDGES_EVENT_SCHEDULE/);
-  assert.doesNotMatch(job, /steps\.central-time\.outputs\.should_run == 'true'/);
-  assert.match(job, /github\.event_name == 'workflow_dispatch'/);
-  assert.match(job, /vars\.CLAWSWEEPER_PROOF_NUDGES_SCHEDULED == '1'/);
-  assert.match(job, /vars\.CLAWSWEEPER_BOT_PROOF_SCHEDULED == '1'/);
-  assert.match(job, /vars\.CLAWSWEEPER_PROOF_NUDGES_EXECUTE == '1'/);
-  assert.match(job, /vars\.CLAWSWEEPER_BOT_PROOF_EXECUTE == '1'/);
-  assert.match(
-    job,
-    /github\.event_name == 'schedule' && \(vars\.CLAWSWEEPER_PROOF_NUDGES_SCHEDULED == '1' \|\| vars\.CLAWSWEEPER_BOT_PROOF_SCHEDULED == '1'\)/,
-  );
-  assert.match(job, /TARGET_REPO_INPUT:/);
-  assert.match(job, /target_repo must be owner\/repo/);
-  assert.match(job, /PROOF_NUDGES_ITEM_NUMBERS:/);
-  assert.match(job, /item_numbers must be a comma-separated list/);
-  assert.match(job, /PROOF_NUDGES_LIMIT:/);
-  assert.match(job, /PROOF_NUDGES_MIN_AGE_DAYS:/);
-  assert.match(job, /PROOF_NUDGES_COOLDOWN_DAYS:/);
-  assert.match(job, /permission-pull-requests: write/);
-  assert.match(
-    job,
-    /numeric_input in PROOF_NUDGES_LIMIT PROOF_NUDGES_MIN_AGE_DAYS PROOF_NUDGES_COOLDOWN_DAYS/,
-  );
-  assert.match(job, /execute_arg=\(\)/);
-  assert.match(job, /if \[ "\$PROOF_NUDGES_EXECUTE" = "true" \]/);
-  assert.match(job, /pnpm run proof-nudges/);
-  assert.match(job, /vars\.CLAWSWEEPER_PROOF_NUDGES_LIMIT/);
-});
-
-test("read-only checkout mode restores file modes and leaves git metadata writable", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const target = join(root, "target");
-    const nested = join(target, "src");
-    const gitDir = join(target, ".git");
-    mkdirSync(nested, { recursive: true });
-    mkdirSync(gitDir, { recursive: true });
-    const sourceFile = join(nested, "app.ts");
-    const executableFile = join(target, "tool.sh");
-    const gitConfig = join(gitDir, "config");
-    writeFileSync(sourceFile, "export const value = 1;\n");
-    writeFileSync(executableFile, "#!/bin/sh\n");
-    writeFileSync(gitConfig, "[core]\n");
-    chmodSync(target, 0o755);
-    chmodSync(nested, 0o750);
-    chmodSync(sourceFile, 0o640);
-    chmodSync(executableFile, 0o755);
-    chmodSync(gitDir, 0o700);
-    chmodSync(gitConfig, 0o600);
-
-    const snapshots = makeTreeReadOnlyForTest(target);
-    assert.equal(statSync(target).mode & 0o777, 0o555);
-    assert.equal(statSync(nested).mode & 0o777, 0o555);
-    assert.equal(statSync(sourceFile).mode & 0o777, 0o444);
-    assert.equal(statSync(executableFile).mode & 0o777, 0o555);
-    assert.equal(statSync(gitDir).mode & 0o777, 0o700);
-    assert.equal(statSync(gitConfig).mode & 0o777, 0o600);
-
-    restoreTreeModesForTest(snapshots);
-    assert.equal(statSync(target).mode & 0o777, 0o755);
-    assert.equal(statSync(nested).mode & 0o777, 0o750);
-    assert.equal(statSync(sourceFile).mode & 0o777, 0o640);
-    assert.equal(statSync(executableFile).mode & 0o777, 0o755);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
-
-test("event review completion removes ClawSweeper eyes reaction", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const block = workflow.slice(
-    workflow.indexOf("- name: React to target item completion"),
-    workflow.indexOf("\n\n  plan:"),
-  );
-
-  assert.match(block, /-f content="\+1"/);
-  assert.match(block, /-f content="eyes"/);
-  assert.match(block, /repos\/\$TARGET_REPO\/issues\/\$ITEM_NUMBER\/reactions\/\$reaction_id/);
-  assert.match(block, /"openclaw-clawsweeper\[bot\]"/);
-  assert.doesNotMatch(block, /issues\/comments\/\$ITEM_NUMBER\/reactions/);
-});
-
-test("event re-review status lets the durable queue reconcile interruptions", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const block = workflow.slice(
-    workflow.indexOf("- name: Mark re-review complete"),
-    workflow.indexOf("- name: Commit event comment router ledger"),
-  );
-
-  assert.match(block, /\[ "\$REVIEW_OUTCOME" = "cancelled" \]/);
-  assert.match(block, /state="Interrupted"/);
-  assert.match(block, /The exact-review queue will reconcile a newer pending item if one arrived/);
-  assert.doesNotMatch(block, /CAPACITY_OUTCOME/);
-  assert.doesNotMatch(block, /state="Superseded"/);
-});
-
-test("event repair retries wait for active worker capacity", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const block = workflow.slice(
-    workflow.indexOf("- name: Detect waiting event repair dispatches"),
-    workflow.indexOf("- name: Commit event comment router retry ledger"),
-  );
-
-  assert.match(block, /--status waiting,active/);
-  assert.match(block, /--wait-for-capacity/);
-});
-
-test("comment commands keep the router-to-sweep dispatch contract", () => {
-  const routerWorkflow = readFileSync(".github/workflows/repair-comment-router.yml", "utf8");
-  const sweepWorkflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const routerSource = readFileSync("src/repair/comment-router.ts", "utf8");
-
-  assert.match(routerWorkflow, /types:\s*\[clawsweeper_comment\]/);
-  assert.match(routerWorkflow, /pnpm run repair:comment-router/);
-  assert.match(
-    routerWorkflow,
-    /status_comment_id="\$\{\{ github\.event\.client_payload\.status_comment_id \|\| '' \}\}"/,
-  );
-  assert.match(routerWorkflow, /--status-comment-id "\$status_comment_id"/);
-  assert.match(routerSource, /event_type:\s*"clawsweeper_item"/);
-  assert.match(routerSource, /adaptiveReviewBudgetForPullRequest\(command\.target\)/);
-  assert.match(routerSource, /const MAX_MEDIA_PREPROCESSING_TIMEOUT_MS = 480_000/);
-  assert.match(routerSource, /media_proof_timeout_ms: reviewBudget\.mediaProofTimeoutMs/);
-  assert.match(routerSource, /reviewBudget\.codexTimeoutMs \+ MAX_MEDIA_PREPROCESSING_TIMEOUT_MS/);
-  assert.doesNotMatch(
-    routerSource,
-    /reviewBudget\.codexTimeoutMs \+ reviewBudget\.mediaProofTimeoutMs/,
-  );
-  assert.match(routerSource, /`codex_timeout_ms=\$\{fallbackCodexTimeoutMs\}`/);
-  assert.match(sweepWorkflow, /types:\s*\[clawsweeper_item,\s*clawsweeper_target_sweep\]/);
-  assert.doesNotMatch(sweepWorkflow, /types:\s*\[[^\]]*clawsweeper_comment/);
-});
-
-test("comment router prunes bare ack comments after updating shared automerge status", () => {
-  const routerSource = readFileSync("src/repair/comment-router.ts", "utf8");
-  const postComment = routerSource.slice(
-    routerSource.indexOf("function postComment("),
-    routerSource.indexOf("\nfunction findExistingCommandStatusComment"),
-  );
-
-  assert.match(postComment, /const existingStatus = findExistingCommandStatusComment\(command\);/);
-  assert.match(postComment, /const precreated = findPrecreatedCommandStatusComment\(command\);/);
-  assert.match(postComment, /const existing = existingStatus \?\? precreated;/);
-  assert.match(
-    postComment,
-    /if \(existingStatus && precreatedId > 0 && precreatedId !== existingId\)/,
-  );
-  assert.match(postComment, /issues\/comments\/\$\{precreatedId\}/);
-  assert.match(postComment, /"DELETE"/);
-  assert.match(postComment, /pruned_ack_comment_id: String\(precreatedId\)/);
-});
-
-test("manual exact-item review dispatches avoid broad review concurrency", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-
-  assert.match(
-    workflow,
-    /github\.event_name == 'workflow_dispatch' && \(github\.event\.inputs\.item_number != '' \|\| github\.event\.inputs\.item_numbers != ''\)\) && format\('clawsweeper-intake-exact-\{0\}'/,
-  );
-  assert.doesNotMatch(
-    workflow,
-    /github\.event_name == 'workflow_dispatch' && github\.event\.inputs\.hot_intake == 'true' && \(github\.event\.inputs\.item_number != '' \|\| github\.event\.inputs\.item_numbers != ''\)\) && format\('clawsweeper-intake-exact-\{0\}'/,
-  );
-});
-
-test("sweep workflow publishes target-scoped state paths", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-
-  assert.match(workflow, /target_slug="\$TARGET_REPO"/);
-  assert.match(workflow, /--path "records\/\$\{target_slug\}"/);
-  assert.match(workflow, /--path "results\/sweep-status\/\$\{target_slug\}\.json"/);
-  assert.doesNotMatch(workflow, /--path records\s*\\/);
-  assert.doesNotMatch(workflow, /--path results\/sweep-status\s*\\/);
-});
-
-test("sweep workflow schedules cursor-based PR comment sync batches", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-
-  assert.match(workflow, /cron: "6,21,36,51 \* \* \* \*"/);
-  assert.doesNotMatch(workflow, /apply_sync_open_pr_batch:/);
-  assert.match(
-    workflow,
-    /sync_batch_size="\$\{\{ github\.event_name == 'workflow_dispatch' && github\.event\.inputs\.apply_limit \|\| '25' \}\}"/,
-  );
-  assert.match(workflow, /\$item_numbers" = "__cursor__"/);
-  assert.match(workflow, /comment-sync-batch/);
-  assert.match(workflow, /write-comment-sync-cursor/);
-  assert.match(workflow, /results\/comment-sync-cursors\/\$\{target_slug\}\.json/);
-  assert.match(workflow, /APPLY_SYNC_OPEN_PR_BATCH/);
-  assert.match(workflow, /github\.event\.schedule == '6,21,36,51 \* \* \* \*'/);
-});
-
-test("sweep target checkouts retry without cached references", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const checkoutBlocks =
-    workflow.match(/- name: Check out target repository[\s\S]*?rev-parse --short HEAD/g) ?? [];
-
-  assert.equal(checkoutBlocks.length, 2);
-  for (const block of checkoutBlocks) {
-    assert.match(block, /Cached target repository fetch failed; rebuilding cache/);
-    assert.match(block, /Cached target checkout failed; retrying without cache reference/);
-    assert.match(block, /rm -rf "\$checkout_dir" "\$cache_dir"/);
-    assert.match(
-      block,
-      /git clone --filter=blob:none --branch "\$target_branch" --single-branch "\$url" "\$checkout_dir"/,
-    );
-  }
-});
-
-test("target sweep runs count as background review capacity", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const capacityBlock = workflow.slice(
-    workflow.indexOf("active_sweep_background_workers()"),
-    workflow.indexOf(
-      'active_critical_workers="$',
-      workflow.indexOf("active_sweep_background_workers()"),
-    ),
-  );
-
-  assert.match(workflow, /Review hot target repo/);
-  assert.match(capacityBlock, /startswith\("Review target repo "\)/);
-  assert.match(capacityBlock, /startswith\("Review hot target repo "\)/);
-  assert.match(capacityBlock, /Review\\ hot\\ target\\ repo/);
-});
-
-test("target hot sweep dispatches honor shard cap payload", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const modeBlock = workflow.slice(
-    workflow.indexOf("- id: mode"),
-    workflow.indexOf("\n      - id: select"),
-  );
-
-  assert.match(modeBlock, /elif \[ "\$hot_intake" = "true" \]; then/);
-  assert.match(
-    modeBlock,
-    /shard_count="\$\{\{ github\.event\.client_payload\.shard_count \|\| '' \}\}"/,
-  );
-  assert.match(modeBlock, /shard_count="\$hot_intake_shards"/);
-});
-
-test("review git info follows checked-out target branch", () => {
-  const source = readFileSync("src/clawsweeper.ts", "utf8");
-
-  assert.match(source, /function reviewTargetBranch/);
-  assert.match(source, /rev-parse", "--abbrev-ref", "HEAD"/);
-  assert.match(source, /refs\/remotes\/origin\/\$\{targetBranch\}/);
-});
-
-test("sweep workflow_dispatch input count stays under GitHub limit", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8").replace(/\r\n/g, "\n");
-  const inputBlock = workflow.slice(
-    workflow.indexOf("  workflow_dispatch:\n    inputs:"),
-    workflow.indexOf("\n  schedule:"),
-  );
-  const inputNames = [...inputBlock.matchAll(/^      [A-Za-z0-9_]+:/gm)];
-
-  assert.ok(inputNames.length <= 25, `workflow_dispatch has ${inputNames.length} inputs`);
-});
-
-test("sweep review continuations stay workflow-dispatch compatible", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8").replace(/\r\n/g, "\n");
-  const continueBlock = workflow.slice(
-    workflow.indexOf("- name: Continue sweep"),
-    workflow.indexOf("\n\n  recover-review-failures:"),
-  );
-  const recoveryBlock = workflow.slice(
-    workflow.indexOf("args=(\n            workflow run sweep.yml"),
-    workflow.indexOf("\n\n  audit-dashboard:"),
-  );
-
-  for (const block of [continueBlock, recoveryBlock]) {
-    assert.match(block, /-f target_repo="\$\{\{ needs\.plan\.outputs\.target_repo \}\}"/);
-    assert.match(block, /-f target_branch="\$\{\{ needs\.plan\.outputs\.target_branch \}\}"/);
-  }
-});
-
-test("target sweep dispatches preserve disabled ClawHub guard", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const planHeader = workflow.slice(
-    workflow.indexOf("\n  plan:"),
-    workflow.indexOf("\n    runs-on:", workflow.indexOf("\n  plan:")),
-  );
-
-  assert.match(planHeader, /github\.event\.action == 'clawsweeper_target_sweep'/);
-  assert.match(
-    planHeader,
-    /github\.event_name == 'repository_dispatch' && github\.event\.client_payload\.target_repo == 'openclaw\/clawhub' && vars\.CLAWSWEEPER_ENABLE_CLAWHUB != '1'/,
-  );
-});
-
-test("sweep planning-started status publish is bounded", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const block = workflow.slice(
-    workflow.indexOf("- name: Publish planning-started status"),
-    workflow.indexOf("- id: mode"),
-  );
-
-  assert.match(block, /timeout 20s pnpm run repair:publish-main/);
-  assert.match(block, /Skipped slow planning-started dashboard publish/);
-});
-
-test("review capacity probes use REST actions run listing", () => {
-  const sweepWorkflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const sweepBlock = sweepWorkflow.slice(
-    sweepWorkflow.indexOf("- id: mode"),
-    sweepWorkflow.indexOf("- id: select"),
-  );
-  const commitWorkflow = readFileSync(".github/workflows/commit-review.yml", "utf8");
-  const commitBlock = commitWorkflow.slice(
-    commitWorkflow.indexOf("- name: Select commits"),
-    commitWorkflow.indexOf('if [ "$ENABLED" = "false" ]'),
-  );
-
-  for (const block of [sweepBlock, commitBlock]) {
-    assert.match(block, /active_runs_json\(\)/);
-    assert.match(block, /actions\/runs\?per_page=100/);
-    assert.match(block, /--paginate/);
-    assert.match(block, /status=\$\{run_status\}/);
-    assert.match(block, /workflowName:\.name/);
-    assert.match(block, /displayTitle:\.display_title/);
-    assert.match(block, /createdAt:\.created_at/);
-    assert.match(block, /updatedAt:\.updated_at/);
-    assert.match(block, /STALE_QUEUED_CUTOFF/);
-    assert.doesNotMatch(block, /gh run list/);
-    assert.match(block, /gh run view/);
-  }
-});
-
-test("background review capacity reserves expanding matrices and caps broad manual input", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const modeBlock = workflow.slice(
-    workflow.indexOf("- id: mode"),
-    workflow.indexOf("- id: select"),
-  );
-  const commitWorkflow = readFileSync(".github/workflows/commit-review.yml", "utf8");
-  const commitBlock = commitWorkflow.slice(
-    commitWorkflow.indexOf("- name: Select commits"),
-    commitWorkflow.indexOf('if [ "$ENABLED" = "false" ]'),
-  );
-
-  assert.match(modeBlock, /limit review_shards\.hot_intake_default/);
-  assert.match(modeBlock, /limit review_shards\.normal_default/);
-  assert.match(modeBlock, /STALE_QUEUED_CUTOFF/);
-  assert.match(modeBlock, /updatedAt:\.updated_at/);
-  assert.match(modeBlock, /lane_shard_cap="\$normal_shards"/);
-  assert.match(modeBlock, /lane_shard_cap="\$hot_intake_shards"/);
-  assert.match(modeBlock, /Capping broad background review shards/);
-  assert.match(commitBlock, /limit review_shards\.hot_intake_default/);
-  assert.match(commitBlock, /limit review_shards\.normal_default/);
-  assert.match(commitBlock, /STALE_QUEUED_CUTOFF/);
-  assert.match(commitBlock, /updatedAt:\.updated_at/);
-});
-
-test("scheduled normal review keeps workers warm with multi-item shards", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const modeBlock = workflow.slice(
-    workflow.indexOf("- id: mode"),
-    workflow.indexOf("- id: select"),
-  );
-
-  assert.match(
-    modeBlock,
-    /if \[ "\$\{\{ github\.event_name \}\}" = "schedule" \]; then\s+batch_size="3"/,
-  );
-});
-
-test("sweep event reviews and target fanout avoid storm amplification", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const legacyIntakeBlock = workflow.slice(
-    workflow.indexOf("legacy-event-queue-intake:"),
-    workflow.indexOf("event-review-apply:"),
-  );
-  const eventBlock = workflow.slice(
-    workflow.indexOf("event-review-apply:"),
-    workflow.indexOf("target-fanout:"),
-  );
-  const fanoutBlock = workflow.slice(workflow.indexOf("target-fanout:"), workflow.indexOf("plan:"));
-
-  assert.match(eventBlock, /concurrency:/);
-  assert.match(
-    eventBlock,
-    /clawsweeper-event-review-\$\{\{ github\.event\.client_payload\.target_repo/,
-  );
-  assert.match(
-    eventBlock,
-    /group: clawsweeper-event-review-\$\{\{ github\.event\.client_payload\.target_repo \|\| 'openclaw\/openclaw' \}\}-\$\{\{ github\.event\.client_payload\.item_number/,
-  );
-  assert.match(eventBlock, /queue_lease_id != ''/);
-  assert.match(eventBlock, /cancel-in-progress: false/);
-  assert.match(legacyIntakeBlock, /legacy-event-queue-intake:/);
-  assert.match(legacyIntakeBlock, /\/internal\/exact-review\/enqueue/);
-  assert.match(
-    fanoutBlock,
-    /FANOUT_LIMIT: \$\{\{ github\.event\.schedule == '41 \* \* \* \*' && '6' \|\| \(github\.event\.schedule == '37 \*\/6 \* \* \*' && '12' \|\| '6'\) \}\}/,
-  );
-});
-
-test("setup-state defaults to an auth-safe shallow checkout", () => {
-  const action = readFileSync(".github/actions/setup-state/action.yml", "utf8");
-  const filterBlock = action.slice(action.indexOf("filter:"), action.indexOf("fetch-depth:"));
-  const fetchDepthBlock = action.slice(action.indexOf("fetch-depth:"), action.indexOf("runs:"));
-
-  assert.match(filterBlock, /default: ""/);
-  assert.doesNotMatch(filterBlock, /default: blob:none/);
-  assert.match(action, /filter: \$\{\{ inputs\.filter \}\}/);
-  assert.match(fetchDepthBlock, /default: "1"/);
-  assert.doesNotMatch(fetchDepthBlock, /default: "0"/);
-  assert.match(action, /fetch-depth: \$\{\{ inputs\.fetch-depth \}\}/);
-  assert.match(action, /sparse-checkout: \$\{\{ inputs\.sparse-checkout \}\}/);
-  assert.doesNotMatch(action, /state-repository:/);
-  assert.doesNotMatch(action, /state-ref:/);
-  assert.match(action, /repository: openclaw\/clawsweeper-state/);
-  assert.match(action, /ref: state/);
-});
-
-test("sweep exact event reviews consume adaptive Codex timeout payload", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const resolveBlock = workflow.slice(
-    workflow.indexOf("- name: Resolve event payload"),
-    workflow.indexOf("- name: Create target read token"),
-  );
-  const reviewBlock = workflow.slice(
-    workflow.indexOf("- name: Review exact event item"),
-    workflow.indexOf("- name: Create state token"),
-  );
-
-  assert.match(
-    resolveBlock,
-    /ADAPTIVE_CODEX_TIMEOUT_MS: \$\{\{ github\.event\.client_payload\.codex_timeout_ms \|\| '' \}\}/,
-  );
-  assert.match(
-    resolveBlock,
-    /CONFIGURED_CODEX_TIMEOUT_MS: \$\{\{ vars\.CLAWSWEEPER_CODEX_TIMEOUT_MS \|\| '1200000' \}\}/,
-  );
-  assert.match(
-    resolveBlock,
-    /MEDIA_PROOF_TIMEOUT_MS: \$\{\{ github\.event\.client_payload\.media_proof_timeout_ms \|\| '0' \}\}/,
-  );
-  assert.match(resolveBlock, /Ignoring invalid adaptive codex_timeout_ms payload/);
-  assert.match(
-    resolveBlock,
-    /configured_codex_timeout_ms="\$\(\(10#\$configured_codex_timeout_ms\)\)"/,
-  );
-  assert.match(
-    resolveBlock,
-    /adaptive_codex_timeout_ms="\$\(\(10#\$adaptive_codex_timeout_ms\)\)"/,
-  );
-  assert.match(resolveBlock, /media_proof_timeout_ms="\$\(\(10#\$media_proof_timeout_ms\)\)"/);
-  assert.match(resolveBlock, /\[ "\$media_proof_timeout_ms" -gt 480000 \]/);
-  assert.match(resolveBlock, /\[ "\$adaptive_codex_timeout_ms" -lt 600000 \]/);
-  assert.match(resolveBlock, /\[ "\$adaptive_codex_timeout_ms" -gt 1800000 \]/);
-  assert.match(resolveBlock, /\[ "\$adaptive_codex_timeout_ms" -gt "\$codex_timeout_ms" \]/);
-  assert.match(resolveBlock, /echo "codex_timeout_ms=\$codex_timeout_ms"/);
-  assert.match(resolveBlock, /echo "media_proof_timeout_ms=\$media_proof_timeout_ms"/);
-  assert.match(
-    reviewBlock,
-    /codex_timeout_ms="\$\{\{ steps\.target\.outputs\.codex_timeout_ms \}\}"/,
-  );
-  assert.match(reviewBlock, /media_preprocessing_reserve_seconds=480/);
-  assert.match(
-    reviewBlock,
-    /review_timeout_seconds=\$\(\(codex_timeout_seconds \+ media_preprocessing_reserve_seconds \+ 180\)\)/,
-  );
-  assert.match(reviewBlock, /detected media allowance \$\{media_proof_timeout_seconds\}s/);
-  assert.doesNotMatch(reviewBlock, /review_timeout_seconds=.*media_proof_timeout_seconds/);
-  assert.match(reviewBlock, /timeout --kill-after=30s "\$\{review_timeout_seconds\}s"/);
-  assert.match(reviewBlock, /--codex-timeout-ms "\$codex_timeout_ms"/);
-  assert.doesNotMatch(reviewBlock, /timeout --kill-after=30s 12m/);
-  assert.doesNotMatch(reviewBlock, /--codex-timeout-ms 600000/);
-});
-
-test("sweep exact event reviews preserve the configured fallback without an adaptive payload", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  const resolveBlock = workflow.slice(
-    workflow.indexOf("- name: Resolve event payload"),
-    workflow.indexOf("- name: Create target read token"),
-  );
-
-  assert.match(
-    resolveBlock,
-    /CONFIGURED_CODEX_TIMEOUT_MS: \$\{\{ vars\.CLAWSWEEPER_CODEX_TIMEOUT_MS \|\| '1200000' \}\}/,
-  );
-  assert.match(resolveBlock, /codex_timeout_ms="\$configured_codex_timeout_ms"/);
-  assert.match(resolveBlock, /\[ "\$adaptive_codex_timeout_ms" -gt "\$codex_timeout_ms" \]/);
-});
-
-test("github activity workflow scopes cancellation to matching item activity", () => {
-  const workflow = readFileSync(".github/workflows/github-activity.yml", "utf8");
-  const concurrencyBlock = workflow.slice(
-    workflow.indexOf("concurrency:"),
-    workflow.indexOf("jobs:"),
-  );
-
-  assert.match(concurrencyBlock, /group: >-/);
-  assert.match(
-    concurrencyBlock,
-    /github-activity-\$\{\{ github\.event\.client_payload\.activity\.repo/,
-  );
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.target_repo/);
-  assert.match(concurrencyBlock, /github\.event\.repository\.full_name/);
-  assert.match(concurrencyBlock, /github\.event_name == 'workflow_run'/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.event_name/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.type/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.action/);
-  assert.match(concurrencyBlock, /github\.event\.action/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.comment_id/);
-  assert.match(concurrencyBlock, /github\.event\.comment\.id/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.review\.id/);
-  assert.match(concurrencyBlock, /github\.event\.review\.id/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.pull_request\.number/);
-  assert.match(concurrencyBlock, /github\.event\.pull_request\.number/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.issue\.number/);
-  assert.match(concurrencyBlock, /github\.event\.issue\.number/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.subject\.number/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.label\.name/);
-  assert.match(concurrencyBlock, /github\.event\.label\.name/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.assignee\.login/);
-  assert.match(concurrencyBlock, /github\.event\.assignee\.login/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.delivery_id/);
-  assert.match(concurrencyBlock, /github\.event\.client_payload\.activity\.idempotency_key/);
-  assert.match(workflow, /Check core API budget/);
-  assert.match(workflow, /CLAWSWEEPER_MIN_CORE_REMAINING/);
-  assert.match(workflow, /contents: write/);
-  assert.doesNotMatch(workflow, /Dispatch spam comment intake candidates/);
-  assert.match(workflow, /Dispatch spam scan candidate/);
-  assert.match(workflow, /repair:spam-comment-intake -- --write-report/);
-  assert.doesNotMatch(workflow, /gh api "repos\/\$\{GITHUB_REPOSITORY\}\/dispatches"/);
-  assert.match(concurrencyBlock, /cancel-in-progress: true/);
-  assert.match(workflow, /runs-on: ubuntu-24\.04/);
-  assert.doesNotMatch(workflow, /runs-on: blacksmith-/);
-  assert.doesNotMatch(
-    concurrencyBlock,
-    /group: github-activity-\$\{\{ github\.event_name \}\}-\$\{\{ github\.run_id \}\}/,
-  );
-  assert.doesNotMatch(concurrencyBlock, /workflow-run' \|\| 'activity'/);
 });
 
 test("spam comment intake coalesces duplicate comment deliveries", () => {
